+++ test/bigtest/helpers/PluginHarness.js
@@ -4,9 +4,9 @@ import { Pluggable } from '@folio/stripes/core';
 
 const PluginHarness = () => (
   <Pluggable
-    aria-haspopup="true"
     type="quick-marc"
-    baseRoute="/dummy"
+    basePath="/dummy"
+    onClose={() => {}}
   >
     <span data-test-no-plugin-available>No plugin available!</span>
   </Pluggable>