+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## (in progress)
 
 ### Stories
+* [UIQC-1](https://issues.folio.org/browse/UIQM-1) Open quickMARC editor from instance record
 * [UIQC-10](https://issues.folio.org/browse/UIQM-10) Project Setup: ui-quick-marc
 
 ### Bug fixes